Normalize depth limits
Before this commit you could set a min limit that was bigger than the
max limit and vice versa, which doesn't make sense and will make the
editor empty. Now, if you set a min limit that's bigger than the max
limit, the max limit will be set to the same value, and vice versa.