Use visibility vectors in filtering
Previously, any update to a filter would require the entire list of
entries be refiltered, with each entry being checked with every filter.
Doing so is wasteful since we can easily store the outcome of the
filters and only recompute the filter that was changed. That is now
being done here