Add list of filtered indices to editor state
This is preparing for making changes in the editor. To know which entry
 a change has been made to, you need to be able to correlate the line
in the editor with the entry. Keeping the list of filtered indices
allows that.

Additionally, this commit changes the sorting to a sort to indices
type thing. This should be more efficient since it's no longer moving
around strings when sorting